# src/app/api/v1/executions.py
from enum import StrEnum
from typing import Annotated, Any, Literal
from uuid import UUID

from fastapi import APIRouter, Depends, HTTPException, Query, Request, status
from fastcrud import PaginatedListResponse, compute_offset, paginated_response
from sqlalchemy.ext.asyncio import AsyncSession

from ...api.dependencies import get_current_user
from ...core.config import settings
from ...core.db.database import async_get_db
from ...core.exceptions.http_exceptions import NotFoundException
from ...core.ledger.service import execution_ledger_service
from ...core.openapi import (
    RESPONSES_NOT_FOUND,
    RESPONSES_SERVER_ERROR,
    RESPONSES_SERVICE_UNAVAILABLE,
    authenticated_responses,
    merge_responses,
)
from ...core.orchestration.service import (
    cancel_scheduler_session_for_execution,
    enrich_execution_dim_rest_urls,
    prepare_execution as orchestration_prepare_execution,
    read_execution_ledger_snapshot,
)
from ...core.utils import queue
from ...crud.crud_daliuge_deployment_profile import crud_daliuge_deployment_profile
from ...crud.crud_execution_record import crud_batch_execution_records
from ...models.ledger import ExecutionStatus
from ...schemas.ledger import (
    BatchExecutionRecordCreate,
    BatchExecutionRecordListItem,
    BatchExecutionRecordRead,
    BatchExecutionRecordUpdate,
    BatchExecutionStatusResponse,
    BatchExecutionSummary,
    ExecuteAcceptedResponse,
    ExecuteRequest,
    ExecutionLedgerSnapshot,
    PrepareExecutionRequest,
    PrepareExecutionResponse,
)


class ExecutionSortField(StrEnum):
    created_at = "created_at"
    updated_at = "updated_at"
    started_at = "started_at"
    completed_at = "completed_at"
    status = "status"

router = APIRouter(prefix="/executions", tags=["executions"])


@router.post(
    "/prepare",
    response_model=PrepareExecutionResponse,
    summary="Prepare execution",
    responses=authenticated_responses(),
)
async def prepare_execution(
    request: Request,
    body: PrepareExecutionRequest,
    current_user: Annotated[dict, Depends(get_current_user)],
    db: Annotated[AsyncSession, Depends(async_get_db)],
) -> dict[str, Any]:
    """Validate sources and preview datasets without creating an execution record."""
    return await orchestration_prepare_execution(
        db=db,
        project_module=body.project_module,
        sources=body.sources,
    )


@router.get(
    "",
    response_model=PaginatedListResponse[BatchExecutionRecordListItem],
    summary="List executions",
    responses=merge_responses(
        {
            status.HTTP_200_OK: {
                "description": "Paginated batch execution records",
            },
        },
    ),
)
async def list_executions(
    request: Request,
    db: Annotated[AsyncSession, Depends(async_get_db)],
    page: Annotated[int, Query(ge=1, description="Page number (1-based)")] = 1,
    items_per_page: Annotated[int, Query(ge=1, le=100, description="Items per page")] = 10,
    project_module: Annotated[
        str | None, Query(description="Filter by project module identifier")
    ] = None,
    execution_status: Annotated[
        ExecutionStatus | None, Query(description="Filter by execution status", alias="status")
    ] = None,
    sort_by: Annotated[ExecutionSortField, Query(description="Column to sort by")] = ExecutionSortField.created_at,
    order: Annotated[Literal["asc", "desc"], Query(description="Sort direction")] = "desc",
) -> dict[str, Any]:
    """List batch execution records with optional filters and pagination."""
    filters: dict[str, Any] = {}
    if project_module:
        filters["project_module"] = project_module
    if execution_status:
        filters["status"] = execution_status

    executions_data = await crud_batch_execution_records.get_multi(
        db=db,
        offset=compute_offset(page, items_per_page),
        limit=items_per_page,
        schema_to_select=BatchExecutionRecordListItem,  # type: ignore[arg-type]
        sort_columns=[sort_by.value],
        sort_orders=[order],
        **filters,
    )

    return paginated_response(
        crud_data=executions_data,  # type: ignore[arg-type]
        page=page,
        items_per_page=items_per_page,
    )


@router.post(
    "",
    response_model=BatchExecutionRecordRead,
    status_code=status.HTTP_201_CREATED,
    summary="Create execution",
    responses=authenticated_responses(RESPONSES_NOT_FOUND),
)
async def create_execution(
    request: Request,
    execution_data: BatchExecutionRecordCreate,
    current_user: Annotated[dict, Depends(get_current_user)],
    db: Annotated[AsyncSession, Depends(async_get_db)],
) -> dict[str, Any]:
    """Create a batch execution record in the ledger."""
    deployment_profile_id: UUID | None = None
    if execution_data.deployment_profile_name:
        profile = await crud_daliuge_deployment_profile.get(
            db=db,
            name=execution_data.deployment_profile_name,
        )
        if profile is None:
            raise NotFoundException(
                f"Deployment profile {execution_data.deployment_profile_name} not found"
            )
        deployment_profile_id = profile.get("uuid")

    return await execution_ledger_service.create_execution(
        db=db,
        project_module=execution_data.project_module,
        sources=[s.model_dump() for s in execution_data.sources],
        archive_name=execution_data.archive_name,
        deployment_profile_id=deployment_profile_id,
        created_by_id=current_user.get("id"),
    )


@router.get(
    "/{execution_id}",
    response_model=BatchExecutionRecordRead,
    summary="Get execution",
    responses=merge_responses(RESPONSES_NOT_FOUND),
)
async def get_execution(
    request: Request,
    execution_id: UUID,
    db: Annotated[AsyncSession, Depends(async_get_db)],
) -> dict[str, Any]:
    """Return a batch execution record by UUID."""
    execution = await crud_batch_execution_records.get(
        db=db, uuid=execution_id, schema_to_select=BatchExecutionRecordRead
    )
    if execution is None:
        raise NotFoundException(f"Execution {execution_id} not found")
    row = dict(execution)
    row.update(await enrich_execution_dim_rest_urls(db, row))
    return row


@router.get(
    "/{execution_id}/ledger-snapshot",
    response_model=ExecutionLedgerSnapshot,
    summary="Get ledger snapshot",
    responses=merge_responses(RESPONSES_NOT_FOUND),
)
async def get_execution_ledger_snapshot(
    request: Request,
    execution_id: UUID,
    db: Annotated[AsyncSession, Depends(async_get_db)],
) -> dict[str, Any]:
    """Return a compact execution view for operators and workflow correlation."""
    execution = await crud_batch_execution_records.get(
        db=db, uuid=execution_id, schema_to_select=BatchExecutionRecordRead
    )
    if execution is None:
        raise NotFoundException(f"Execution {execution_id} not found")
    return await read_execution_ledger_snapshot(db=db, execution_id=execution_id)


_SCANCEL_ELIGIBLE_STATUSES: frozenset[ExecutionStatus] = frozenset(
    {ExecutionStatus.AWAITING_SCHEDULER, ExecutionStatus.RUNNING}
)


def _coerce_execution_status(raw: Any) -> ExecutionStatus | None:
    if raw is None:
        return None
    if isinstance(raw, ExecutionStatus):
        return raw
    try:
        return ExecutionStatus(str(raw))
    except ValueError:
        return None


@router.patch(
    "/{execution_id}",
    response_model=BatchExecutionRecordRead,
    summary="Update execution",
    responses=authenticated_responses(RESPONSES_NOT_FOUND),
)
async def update_execution(
    request: Request,
    execution_id: UUID,
    execution_update: BatchExecutionRecordUpdate,
    current_user: Annotated[dict, Depends(get_current_user)],
    db: Annotated[AsyncSession, Depends(async_get_db)],
) -> dict[str, Any]:
    """Update execution status, manifest, or scheduler fields."""
    if execution_update.status == ExecutionStatus.CANCELLED:
        existing = await crud_batch_execution_records.get(
            db=db, uuid=execution_id, schema_to_select=BatchExecutionRecordRead
        )
        if existing is None:
            raise NotFoundException(f"Execution {execution_id} not found")
        current_status_enum = _coerce_execution_status(existing.get("status"))
        if current_status_enum in _SCANCEL_ELIGIBLE_STATUSES:
            await cancel_scheduler_session_for_execution(db=db, execution_id=execution_id)

    return await execution_ledger_service.update_execution_status(
        db=db,
        execution_id=execution_id,
        status=execution_update.status,
        scheduler_job_id=execution_update.scheduler_job_id,
        scheduler_name=execution_update.scheduler_name,
        workflow_manifest=execution_update.workflow_manifest,
        error=execution_update.last_error,
    )


@router.get(
    "/{execution_id}/status",
    response_model=BatchExecutionStatusResponse,
    summary="Get execution status",
    responses=merge_responses(RESPONSES_NOT_FOUND),
)
async def get_execution_status(
    request: Request,
    execution_id: UUID,
    db: Annotated[AsyncSession, Depends(async_get_db)],
) -> dict[str, Any]:
    """Return status-focused fields for an execution."""
    execution = await crud_batch_execution_records.get(
        db=db,
        uuid=execution_id,
        schema_to_select=BatchExecutionRecordRead,
    )
    if execution is None:
        raise NotFoundException(f"Execution {execution_id} not found")
    return BatchExecutionStatusResponse.model_validate(execution).model_dump()


@router.get(
    "/{execution_id}/summary",
    response_model=BatchExecutionSummary,
    summary="Get execution summary",
    responses=merge_responses(RESPONSES_NOT_FOUND),
)
async def get_execution_summary(
    request: Request,
    execution_id: UUID,
    db: Annotated[AsyncSession, Depends(async_get_db)],
) -> dict[str, Any]:
    """Return a summarized view of execution progress and timing."""
    execution = await crud_batch_execution_records.get(
        db=db,
        uuid=execution_id,
        schema_to_select=BatchExecutionRecordRead,
    )
    if execution is None:
        raise NotFoundException(f"Execution {execution_id} not found")
    return BatchExecutionSummary.model_validate(execution).model_dump()


@router.post(
    "/{execution_id}/execute",
    response_model=ExecuteAcceptedResponse,
    status_code=status.HTTP_202_ACCEPTED,
    summary="Execute execution",
    responses=authenticated_responses(
        merge_responses(
            RESPONSES_NOT_FOUND,
            RESPONSES_SERVICE_UNAVAILABLE,
            RESPONSES_SERVER_ERROR,
        ),
    ),
)
async def execute_execution(
    request: Request,
    execution_id: UUID,
    current_user: Annotated[dict, Depends(get_current_user)],
    db: Annotated[AsyncSession, Depends(async_get_db)],
    body: ExecuteRequest | None = None,
) -> dict[str, Any]:
    """Enqueue staging and/or DALiuGE submit for an existing execution record."""
    do_stage = body.do_stage if body else True
    do_submit = body.do_submit if body else True

    execution = await crud_batch_execution_records.get(
        db=db, uuid=execution_id, schema_to_select=BatchExecutionRecordRead
    )
    if execution is None:
        raise NotFoundException(f"Execution {execution_id} not found")

    if queue.pool is None:
        raise HTTPException(status_code=503, detail="Queue is not available")

    job = await queue.pool.enqueue_job(
        "execute_execution_job",
        str(execution_id),
        do_stage=do_stage,
        do_submit=do_submit,
        _queue_name=settings.WORKER_QUEUE_NAME,
    )
    if job is None:
        raise HTTPException(status_code=500, detail="Failed to enqueue execution job")

    return {
        "status": "accepted",
        "execution_id": str(execution_id),
        "job_id": job.job_id,
        "do_stage": do_stage,
        "do_submit": do_submit,
    }